Changed node floats to doubles to preserve precision.  Text.parseNum() also now works at full precision

+++ include/tree.h
@@ -159,7 +159,7 @@ Node* MakeNodeFromTypeArray(TypeArray* thearray, YYLTYPE loc);
 
 Node* MakeNodeFromString(int nodetype, char* mystring, YYLTYPE loc);
 
-Node* MakeNodeFromNumber(int nodetype, float number, YYLTYPE loc);
+Node* MakeNodeFromNumber(int nodetype, double number, YYLTYPE loc);
 
 Node* MakeTwoBranchNode(int nodetype, Node* a, Node* b, YYLTYPE loc);
